pyproject.toml
[tool.poetry]
name = "qprofiler"
version = "0.3.0"
description = "profile tabular datasets, manage automatic validation for new datasets, automatic handling for quality issues."
authors = ["Ezzaldin97 <kingtal632@yahoo.com>"]
repository = "https://github.com/Ezzaldin97/dprofiler"
license = "MIT"
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
polars = {version = "^0.19.0", extras = ["all"]}
pyyaml = "^6.0.1"
pathlib = "^1.0"
ruamel-yaml = "^0.17.32"
colorama = "^0.4.4"
typing-extensions = "^4.7.0"


[tool.poetry.group.test.dependencies]
pytest = "^7.4.0"


[tool.poetry.group.pylint.dependencies]
ruff = "^0.0.287"
black = "^23.3.0"


[tool.poetry.group.dev.dependencies]
ipykernel = "^6.25.2"


[tool.poetry.group.docstring.dependencies]
interrogate = "^1.4.0"

[tool.ruff]
line-length = 88
src = ["qprofiler"]

[tool.ruff.per-file-ignores]
"__init__.py" = ["F401", "E402"]
"qprofiler/tests/*.py" = ["E402", "E712"]

[tool.black]
line-length = 88
target_version = ['py38', 'py39', 'py310']
preview = true
include = 'qprofiler\/.*\.pyi?$'

[tool.interrogate]
ignore-init-method = true
ignore-init-module = true
ignore-magic = true
ignore-semiprivate = true
ignore-private = true
ignore-property-decorators = false
ignore-module = true
ignore-nested-functions = true
ignore-nested-classes = false
ignore-setters = false
fail-under = 95
exclude = ["./qprofiler/tests", "./qprofiler/utils"]
ignore-regex = ["^get$", "^mock_.*", ".*BaseClass.*"]
verbose = 1
quiet = false
whitelist-regex = []
color = true
omit-covered-files = false
generate-badge = "./badges"
badge-format = "svg"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
